# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Enterprise untuk sistem audit log berbasis Blockchain dan Merkle Tree.",
        "title": "AuditChain Gateway API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@auditchain.local"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/recovery/incidents": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil daftar semua insiden tamper berdasarkan client_id (opsional difilter dengan status).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recovery"
                ],
                "summary": "List tamper incidents",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter Status Insiden (OPEN, RESOLVED)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar Insiden Tamper",
                        "schema": {
                            "$ref": "#/definitions/recovery.IncidentListResponse"
                        }
                    },
                    "401": {
                        "description": "Identitas client tidak valid",
                        "schema": {
                            "$ref": "#/definitions/recovery.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Gagal mengambil daftar tamper incident",
                        "schema": {
                            "$ref": "#/definitions/recovery.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/recovery/incidents/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil detail sebuah insiden tamper berdasarkan ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recovery"
                ],
                "summary": "Get tamper incident detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Incident ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail Insiden Tamper",
                        "schema": {
                            "$ref": "#/definitions/recovery.IncidentDetailResponse"
                        }
                    },
                    "401": {
                        "description": "Identitas client tidak valid",
                        "schema": {
                            "$ref": "#/definitions/recovery.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Tamper incident tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/recovery.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Gagal mengambil detail tamper incident",
                        "schema": {
                            "$ref": "#/definitions/recovery.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/recovery/incidents/{id}/candidates": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil daftar snapshot valid yang tersedia di S3 untuk di-recovery pada insiden tertentu.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recovery"
                ],
                "summary": "List candidate snapshots for recovery",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Incident ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar kandidat snapshot",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Identitas client tidak valid",
                        "schema": {
                            "$ref": "#/definitions/recovery.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Insiden tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/recovery.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Gagal mengambil daftar kandidat",
                        "schema": {
                            "$ref": "#/definitions/recovery.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/recovery/incidents/{id}/preflight": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Melakukan pengecekan preflight pada kandidat snapshot untuk memastikan tidak ada konflik sebelum recovery.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recovery"
                ],
                "summary": "Preflight check for recovery",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Incident ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hasil preflight check",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Identitas client tidak valid",
                        "schema": {
                            "$ref": "#/definitions/recovery.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Konflik atau kondisi tidak valid untuk preflight",
                        "schema": {
                            "$ref": "#/definitions/recovery.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Gagal melakukan preflight",
                        "schema": {
                            "$ref": "#/definitions/recovery.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/recovery/requests": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil daftar permintaan (request) recovery yang diajukan.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recovery"
                ],
                "summary": "List recovery requests",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter Status Request",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar recovery request",
                        "schema": {
                            "$ref": "#/definitions/recovery.RequestListResponse"
                        }
                    },
                    "401": {
                        "description": "Identitas client tidak valid",
                        "schema": {
                            "$ref": "#/definitions/recovery.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Gagal mengambil daftar recovery request",
                        "schema": {
                            "$ref": "#/definitions/recovery.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Membuat permintaan baru untuk melakukan recovery data ke snapshot tertentu.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recovery"
                ],
                "summary": "Create recovery request",
                "parameters": [
                    {
                        "description": "Data Request Recovery",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/recovery.CreateRequestInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Recovery request berhasil dibuat",
                        "schema": {
                            "$ref": "#/definitions/recovery.RequestDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Format request tidak valid",
                        "schema": {
                            "$ref": "#/definitions/recovery.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Identitas client tidak valid",
                        "schema": {
                            "$ref": "#/definitions/recovery.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Konflik atau state tidak valid",
                        "schema": {
                            "$ref": "#/definitions/recovery.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Gagal membuat recovery request",
                        "schema": {
                            "$ref": "#/definitions/recovery.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/recovery/requests/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil detail sebuah permintaan (request) recovery berdasarkan ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recovery"
                ],
                "summary": "Get recovery request detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail recovery request",
                        "schema": {
                            "$ref": "#/definitions/recovery.RequestDetailResponse"
                        }
                    },
                    "401": {
                        "description": "Identitas client tidak valid",
                        "schema": {
                            "$ref": "#/definitions/recovery.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Recovery request tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/recovery.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Gagal mengambil detail recovery request",
                        "schema": {
                            "$ref": "#/definitions/recovery.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/recovery/requests/{id}/approve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menyetujui sebuah permintaan recovery (untuk backward compatibility).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recovery"
                ],
                "summary": "Approve recovery request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request berhasil disetujui",
                        "schema": {
                            "$ref": "#/definitions/recovery.RequestDetailResponse"
                        }
                    },
                    "401": {
                        "description": "Identitas client tidak valid",
                        "schema": {
                            "$ref": "#/definitions/recovery.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Request tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/recovery.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "State request tidak sesuai",
                        "schema": {
                            "$ref": "#/definitions/recovery.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/recovery/requests/{id}/execute": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengeksekusi permintaan recovery yang telah disetujui atau siap dieksekusi, memulihkan data dari S3.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recovery"
                ],
                "summary": "Execute recovery request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eksekusi recovery dimulai/berhasil",
                        "schema": {
                            "$ref": "#/definitions/recovery.RequestDetailResponse"
                        }
                    },
                    "401": {
                        "description": "Identitas client tidak valid",
                        "schema": {
                            "$ref": "#/definitions/recovery.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Request tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/recovery.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "State request tidak sesuai",
                        "schema": {
                            "$ref": "#/definitions/recovery.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Gagal mengeksekusi recovery",
                        "schema": {
                            "$ref": "#/definitions/recovery.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/recovery/requests/{id}/reject": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menolak sebuah permintaan recovery.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recovery"
                ],
                "summary": "Reject recovery request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request berhasil ditolak",
                        "schema": {
                            "$ref": "#/definitions/recovery.RequestDetailResponse"
                        }
                    },
                    "401": {
                        "description": "Identitas client tidak valid",
                        "schema": {
                            "$ref": "#/definitions/recovery.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Request tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/recovery.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "State request tidak sesuai",
                        "schema": {
                            "$ref": "#/definitions/recovery.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "recovery.CreateRequestInput": {
            "type": "object",
            "required": [
                "idempotency_key",
                "incident_id",
                "reason",
                "selected_log_id"
            ],
            "properties": {
                "idempotency_key": {
                    "type": "string"
                },
                "incident_id": {
                    "type": "string"
                },
                "reason": {
                    "type": "string",
                    "minLength": 5
                },
                "selected_log_id": {
                    "type": "string"
                }
            }
        },
        "recovery.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Pesan kesalahan"
                }
            }
        },
        "recovery.IncidentDetailResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.TamperIncident"
                }
            }
        },
        "recovery.IncidentListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TamperIncident"
                    }
                }
            }
        },
        "recovery.RequestDetailResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.RecoveryRequest"
                }
            }
        },
        "recovery.RequestListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.RecoveryRequest"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token dengan format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}